+++ src/Extensions/Filters/FormatDateExtension.php
@@ -27,7 +27,7 @@ class FormatDateExtension extends AbstractExtension
     public function getFilters(): array
     {
         return [
-            new TwigFilter('format_date', [$this, 'formatDate'], ['is_safe' => ['html']]),
+            new TwigFilter('format_date', [$this, 'formatDate']),
         ];
     }